created fn to handle text vector
